I made it so NPC ship will move 1 by 1. And they will move their entire speed UNLESS there is a chance of collision. I bult a system where they will move up to the closest empty space next to their target and no further.
For the player, I did some extra checkers that do that too. If you are sailing, and part way through your move you will collide with something, it stops you at the empty space next to the object. There was an issue where if you are next to an object, and had 3 movement, you would skip over the object to a few spaces over. I made checkers that make that acction illegal, and affords the player an opportunity to press a new key to move a different direction